F moved Boost code into approvaltests
